pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jaotarzan", email = "joaovssouza59@gmail.com"},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0", "python-dotenv>=1.1.0", "djangorestframework-simplejwt>=5.5.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
